+++ source/react/library/form/Form.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Alert from '../alert';
-import Button from '../buttons/Button';
+import Button from '../button/Button';
 import {
   componentHasType,
   mapObj,